Version 14.3.0

+++ Classes/ViewHelpers/File/SaveContentToFileViewHelper.php
@@ -19,8 +19,6 @@ namespace YolfTypo3\SavLibraryKickstarter\ViewHelpers\File;
 
 use TYPO3\CMS\Core\Utility\GeneralUtility;
 use TYPO3Fluid\Fluid\Core\ViewHelper\AbstractViewHelper;
-use TYPO3Fluid\Fluid\Core\Rendering\RenderingContextInterface;
-use TYPO3Fluid\Fluid\Core\ViewHelper\Traits\CompileWithRenderStatic;
 use YolfTypo3\SavLibraryKickstarter\Managers\ConfigurationManager;
 
 /**
@@ -37,46 +35,44 @@ use YolfTypo3\SavLibraryKickstarter\Managers\ConfigurationManager;
  *
  * @package SavLibraryKickstarter
  */
-class SaveContentToFileViewHelper extends AbstractViewHelper
+final class SaveContentToFileViewHelper extends AbstractViewHelper
 {
-    use CompileWithRenderStatic;
 
     /**
      * Initializes arguments.
      *
      * @return void
      */
-    public function initializeArguments()
+    public function initializeArguments(): void
     {
-        $this->registerArgument('content', 'string', 'Content to save', true);
+        $this->registerArgument('content', 'string', 'Content to save', false, '');
         $this->registerArgument('extensionKey', 'string', 'Extension key', true);
         $this->registerArgument('fileName', 'string', 'File name', true);
         $this->registerArgument('directory', 'string', 'Directory to create', false, '');
+        $this->registerArgument('doNotCreateIfFileExists', 'bool', 'If true, the file is not created ', false, false);
     }
 
     /**
      * Saves the content into the file
      *
-     * @param array $arguments
-     * @param \Closure $renderChildrenClosure
-     * @param RenderingContextInterface $renderingContext
-     *
      * @return void
      */
-    public static function renderStatic(array $arguments, \Closure $renderChildrenClosure, RenderingContextInterface $renderingContext)
+    public function render(): void
     {
         // Gets the arguments
-        $content = $arguments['content'];
-        $extensionKey = $arguments['extensionKey'];
-        $fileName = $arguments['fileName'];
-        $directory = rtrim($arguments['directory'], '/');
-
+        $content = $this->arguments['content'] ?? '';
+        $extensionKey = $this->arguments['extensionKey'];
+        $fileName = $this->arguments['fileName'];
+        $directory = rtrim($this->arguments['directory'], '/');
+        $doNotCreateIfFileExists = $this->arguments['doNotCreateIfFileExists'];
+        
         // Creates a new directory if needed
         $extensionDirectory = ConfigurationManager::getExtensionDir($extensionKey);
         if (! empty($directory)) {
             GeneralUtility::mkdir_deep($extensionDirectory . $directory .'/');
         }
-
-        GeneralUtility::writeFile($extensionDirectory . $directory . '/' . $fileName, $content);
+        if (! $doNotCreateIfFileExists || ! file_exists($extensionDirectory . $directory . '/' . $fileName)) {
+            GeneralUtility::writeFile($extensionDirectory . $directory . '/' . $fileName, $content);
+        }
     }
 }
